requirements.txt
Django==4.2.1
django-cors-headers==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
requests==2.31.0
requests-oauthlib==1.3.1
twilio==8.13.0
django-allauth
dj-rest-auth
python-dotenv